+++ CefSharp/CefSharp.csproj
@@ -92,6 +92,7 @@
     <Compile Include="CefFileDialogMode.cs" />
     <Compile Include="DefaultResourceHandler.cs" />
     <Compile Include="DragOperationsMask.cs" />
+    <Compile Include="IBitmapFactory.cs" />
     <Compile Include="IDragData.cs" />
     <Compile Include="IFocusHandler.cs" />
     <Compile Include="IDragHandler.cs" />
