+++ CefSharp/IBitmapFactory.cs
@@ -2,14 +2,12 @@
 //
 // Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
 
-using System.Windows.Media.Imaging;
 using CefSharp.Internals;
 
-namespace CefSharp.Wpf
+namespace CefSharp
 {
-	public abstract class WpfBitmapInfo : BitmapInfo
+	public interface IBitmapFactory
 	{
-		public abstract void Invalidate();
-		public abstract BitmapSource CreateBitmap();
+		BitmapInfo CreateBitmap(bool isPopup);
 	}
 }
